data-acquisition/requirements.txt
google-generativeai
PyGithub
requests